refactor: extract session summary widget
Move the SessionSummary widget template and its update logic into a dedicated session_detail submodule. Keep its API restricted to the parent module so SessionDetail only depends on widget(), update(), and template initialization.

Move summary-specific GTK tests next to the implementation and instantiate the template directly. Retain the popover-hosting assertion in SessionDetail tests because it verifies integration between the parent component and the extracted widget.
