+++ src/ui/session_detail/tests.rs
@@ -475,124 +475,6 @@ fn close_inspector_resets_inspector_open() {
 }
 
 #[gtk::test]
-fn session_summary_hides_optional_sections_when_data_is_missing() {
-    let temp_db = tempfile::NamedTempFile::new().expect("temp db");
-    let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
-
-    controller.emit(SessionDetailMsg::SetSession {
-        session: Box::new(build_test_session(None, None, 0, 0, 0)),
-        search_query: None,
-    });
-
-    while gtk::glib::MainContext::default().iteration(false) {}
-
-    let parts = controller.state().get();
-    assert!(!parts.widgets.summary.first_prompt_section.is_visible());
-    assert!(!parts.widgets.summary.tokens_section.is_visible());
-    assert!(
-        parts
-            .widgets
-            .summary
-            .activity_bar
-            .has_css_class("activity-bar")
-    );
-}
-
-#[gtk::test]
-fn session_summary_populates_identity_prompt_and_tokens() {
-    let temp_db = tempfile::NamedTempFile::new().expect("temp db");
-    let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
-
-    controller.emit(SessionDetailMsg::SetSession {
-        session: Box::new(build_test_session(
-            Some("Ship the summary header"),
-            Some(crate::models::TokenUsage {
-                input_tokens: 1200,
-                output_tokens: 300,
-                cache_read_tokens: Some(400),
-                cache_write_tokens: None,
-                reasoning_tokens: Some(50),
-            }),
-            4,
-            2,
-            1,
-        )),
-        search_query: None,
-    });
-
-    while gtk::glib::MainContext::default().iteration(false) {}
-
-    let parts = controller.state().get();
-    assert_eq!(parts.widgets.summary.project_label.label(), "project");
-    assert_eq!(parts.widgets.summary.duration_chip.label(), "2h 14m");
-    assert_eq!(
-        parts.widgets.summary.message_count_chip.label(),
-        "42 messages"
-    );
-    assert_eq!(
-        parts.widgets.summary.ending_status_chip.label(),
-        "Ended cleanly"
-    );
-    assert_eq!(
-        parts.widgets.summary.first_prompt_label.label(),
-        "Ship the summary header"
-    );
-    assert!(
-        parts
-            .widgets
-            .summary
-            .input_value_label
-            .label()
-            .starts_with('1')
-    );
-    assert!(
-        parts
-            .widgets
-            .summary
-            .input_value_label
-            .label()
-            .ends_with("200")
-    );
-    assert_eq!(parts.widgets.summary.output_value_label.label(), "300");
-}
-
-#[gtk::test]
-fn session_summary_uses_conversation_only_when_activity_counts_are_zero() {
-    let temp_db = tempfile::NamedTempFile::new().expect("temp db");
-    let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
-
-    controller.emit(SessionDetailMsg::SetSession {
-        session: Box::new(build_test_session(Some("Only chat"), None, 0, 0, 0)),
-        search_query: None,
-    });
-
-    while gtk::glib::MainContext::default().iteration(false) {}
-
-    let parts = controller.state().get();
-    assert_eq!(
-        parts.widgets.summary.conversation_only_label.label(),
-        "Conversation only"
-    );
-    assert!(!parts.widgets.summary.activity_bar.is_visible());
-}
-
-#[gtk::test]
-fn session_summary_activity_bar_does_not_lock_a_minimum_width_request() {
-    let temp_db = tempfile::NamedTempFile::new().expect("temp db");
-    let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
-
-    controller.emit(SessionDetailMsg::SetSession {
-        session: Box::new(build_test_session(Some("Ship it"), None, 4, 2, 1)),
-        search_query: None,
-    });
-
-    while gtk::glib::MainContext::default().iteration(false) {}
-
-    let parts = controller.state().get();
-    assert_eq!(parts.widgets.summary.activity_bar.width_request(), -1);
-}
-
-#[gtk::test]
 fn session_detail_transcript_list_is_direct_scrolled_window_child() {
     let temp_db = tempfile::NamedTempFile::new().expect("temp db");
     let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
@@ -643,48 +525,6 @@ fn session_detail_content_column_contains_only_transcript_scroller() {
     assert_eq!(
         content.first_child(),
         Some(parts.widgets.transcript_scroller.clone().upcast())
-    );
-}
-
-fn build_error_session_for_css_test() -> Session {
-    let mut session = build_test_session(None, None, 0, 0, 0);
-    session.ending_status = crate::models::SessionEndingStatus::Error;
-    session
-}
-
-#[gtk::test]
-fn session_summary_applies_status_and_activity_css_classes() {
-    let temp_db = tempfile::NamedTempFile::new().expect("temp db");
-    let controller = SessionDetail::builder().launch(temp_db.path().to_path_buf());
-
-    controller.emit(SessionDetailMsg::SetSession {
-        session: Box::new(build_error_session_for_css_test()),
-        search_query: None,
-    });
-
-    while gtk::glib::MainContext::default().iteration(false) {}
-
-    let parts = controller.state().get();
-    assert!(
-        parts
-            .widgets
-            .summary
-            .ending_status_chip
-            .has_css_class("pill")
-    );
-    assert!(
-        parts
-            .widgets
-            .summary
-            .ending_status_chip
-            .has_css_class("ending-failed")
-    );
-    assert!(
-        parts
-            .widgets
-            .summary
-            .activity_bar
-            .has_css_class("activity-bar")
     );
 }
